+++ Changelog.md
@@ -109,6 +109,8 @@ Style Notes
 
 * CI: added sky130 full-chip design.
 
+* CI: added gf180mcu full-chip design.
+
 * CI: added test cases for `gf180mcu_fd_sc_mcu9t5v0` and `gf180mcu_as_sc_mcu7t3v3`.
 
 * Fixed handling of `PAD_CELL_LIBRARY`, now available in config, synthesis and lint.